Separate inline definitions
