+++ component.mk
@@ -1,3 +1,3 @@
-COMPONENT_ADD_INCLUDEDIRS = src
+COMPONENT_ADD_INCLUDEDIRS = include
 COMPONENT_SRCDIRS = src
 COMPONENT_DEPENDS = log driver
